requirements.txt
dwave-ocean-sdk
networkx[default]
matplotlib